replaced MySQL test to mysql-strict-encrypt module
